refactor: Model.Class as the single source of truth + drop ReferenceIdService
Model unification
-----------------
- `Applicant` is now a `Model.Class` in `packages/shared/src/domain/Applicant.ts`
  alongside the branded leaves (FirstName, Email, Phone, NetWorth, ApplicantId).
  It's the single source of truth for FE wire shape, BE handler types, and DB
  schema. Adding/removing a field propagates everywhere.
- `IntakePayload` derives from `Schema.Struct(Applicant.fields).mapFields(Struct.pick(...))`
  instead of duplicating the field list.
- `ApplicantRepo.insert` uses `SqlModel.makeRepository(Applicant, ...)` for the
  generated CRUD machinery (INSERT ... RETURNING * + Model-driven decode of the
  result row). The wrapper service exists only to confine `Redacted.value(...)`
  to a single line — the call site that hands the insert variant to the
  generated repo. Audit-boundary invariant preserved: grep returns one hit.
- Deleted the standalone `packages/backend/src/db/schema/Applicant.ts`.

Model field variant choices
---------------------------
- `id: Model.GeneratedByApp(ApplicantId)` so the id is in `insert.Type` and
  `update.Type` (the SqlModel idColumn constraint requires both). The app
  generates the UUID inside the repo wrapper via
  `Schema.decodeUnknownEffect(ApplicantId)(crypto.randomUUID())`; the
  migration's `DEFAULT gen_random_uuid()` is now a defensive fallback the
  app bypasses.
- `createdAt: Model.GeneratedByDb(Schema.DateTimeUtcFromDate)` — read-only.
  The column's `DEFAULT CURRENT_TIMESTAMP` provides the value; the app never
  supplies it.

Drop ReferenceIdService
-----------------------
The synthetic `ref_<base36>_<6chars>` id was a relic from before the DB
existed — it was thrown away and the client couldn't look it up. The natural
answer to "what's the reference for this submission?" is the row's
`ApplicantId`. `IntakeService.intake` now returns the DB UUID as
`referenceId`. The wire contract is unchanged (Schema.String accepts both
shapes); only the actual value changed.

netWorth: cents → whole dollars
-------------------------------
SqlModel's strict Model-driven row decode caught a real issue: Postgres BIGINT
columns come back as JS strings from node-postgres, and `Schema.Int` doesn't
accept strings. Rather than configure a pg type parser or pollute the wire
schema with a string-accepting union, switched the column to INTEGER and the
schema to whole-dollar values (cap $2B, well within both INT32 and
Number.MAX_SAFE_INTEGER, and well above any realistic "are you accredited"
threshold). Cents-level precision wasn't meaningful for an intake form.

Verified
--------
- `bun --filter @effect-redacted/backend typecheck` clean across all three workspaces.
- `bun run test:unit` 7/7 pass.
- `bun run test:integration` (5 prop runs against testcontainers Postgres) passes.
- Live end-to-end: dev backend with `DATABASE_URL` set, `curl -X POST /intake`
  returns `referenceId: "<uuid>"`, `psql` confirms the row exists with the raw
  `net_worth` integer (no leak).
- Single `Redacted.value(` call site in `packages/backend/src/` — grep-verified.

+++ packages/backend/src/migrations/0001_create_applicants.ts
@@ -9,7 +9,7 @@ export default Effect.gen(function* () {
     last_name TEXT NOT NULL,
     email TEXT NOT NULL,
     phone TEXT NOT NULL,
-    net_worth BIGINT NOT NULL,
+    net_worth INTEGER NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP
   )`
 })
